Revert "Merge pull request #38 from ec-jrc/warmstartfix"
This reverts commit ddaef5a8a2dbee104e7e0a1efc49379e57c3df08, reversing
changes made to 681ea2ccc2848ad55c8de6bf904fa8ee72959cd4.

+++ src/lisflood/hydrological_modules/reservoir.py
@@ -185,7 +185,7 @@ class reservoir(HydroModule):
         option = settings.options
         maskinfo = MaskInfo.instance()
         if option['simulateReservoirs']:
-            InvDtSecDay = float(1/ 86400.)
+            InvDtSecDay = 1 / 86400.
             # InvDtSecDay=self.var.InvDtSec
             # ReservoirInflow = cover(ifthen(defined(self.var.ReservoirSites), upstream(
             # self.var.LddStructuresKinematic, self.var.ChanQ)), scalar(0.0))
